--- Assets/UniGLTF/Runtime/UniGLTF/IO/MeshIO/MeshImporter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d9658f6e46942b146ad7e752c6592401
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
